apply PR comment suggestions (1/2)
