+++ src/app/chat/page.tsx
@@ -1,7 +1,7 @@
 import ChatLayout from '@/components/ChatLayout/ChatLayout';
 import PartnerList from '@/components/PartnerList';
 import ChatBox from '@/components/ChatBox';
-import CurrentPartnerProvider from '@/components/CurrentPartnerProvider';
+import CurrentPartnerProvider from '@/contexts/CurrentPartnerProvider';
 
 import { getPartnerList } from '@/mocks/partner';
 import { PartnerPreviewType } from '@/components/PartnerPreviewChat';
