+++ src/app/layout.tsx
@@ -3,11 +3,9 @@ import '@/app/globals.css';
 import StyledComponentsRegistry from '@/styled-component-registry/registry';
 import GlobalStyles from '@/styled-component-registry/global-styles';
 
-import ThemeProvider from '@/components/ThemeProvider';
+import ThemeProvider from '@/contexts/ThemeProvider';
 
-// todo: create a dark-mode functionality
 import { getThemePreference } from '@/utils/theme';
-
 import { Funnel_Display } from 'next/font/google';
 import type { Metadata } from 'next';
