+++ src/components/PartnerPreviewChat/PartnerPreviewChat.tsx
@@ -4,8 +4,8 @@ import styled from 'styled-components';
 import Avatar from '@/components/Avatar';
 import Interactable from '@/components/Interactable';
 
-import { CurrentPartnerContext } from '@/components/CurrentPartnerProvider';
-import { ChatEnablementContext } from '@/components/ChatLayout/ChatEnablementProvider';
+import { CurrentPartnerContext } from '@/contexts/CurrentPartnerProvider';
+import { ChatEnablementContext } from '@/contexts/ChatEnablementProvider/ChatEnablementProvider';
 
 export type PartnerPreviewType = {
   avatarSrc: string;
