+++ src/contexts/ChatEnablementProvider/index.ts
@@ -1,0 +1,2 @@
+export * from './ChatEnablementProvider';
+export { default } from './ChatEnablementProvider';
